+++ changelog.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Add CDP integration tests to CI (`2da6a4f`)
 - Implement a `CollisionPolicy` rather than a hard-coded gap in `make_draggable()` (`b472e95`)
+- Validate `CollisionPolicy::PushClear::padding` (``)
 
 ## Fixed